CodeBlock Component #2
An issue occurred in production mode with the react-highlight library, specifically the <Highlight/> component, where the syntax highlighting wasn't functioning correctly.

The text remained in the default color, indicating that the expected highlighting wasn't applied.

To resolve this issue, we implemented the following changes:

- Instead of relying on the react-highlight library, we directly imported the original highlight.js library (written in JavaScript) into our custom <CodeBlock/> component.

- We took control of loading the highlight.js library ourselves and ensured that it was fully loaded and executed before displaying the code block.

- Display the codeblock

-Additionally, we improved the flexibility of the <CodeBlock/> component by allowing to choose the font size using a class-based approach.
We can now select from "small", "medium", or "big" font sizes for the code block.
This also sets the stage for future responsiveness by utilizing media queries.

With these adjustments, the <Highlight/> component now functions as intended in both development and production environments, and the syntax highlighting is properly applied to the code blocks.